+++ Sources/TextDiffCore/DiffTypes.swift
