+++ Sources/TextDiffCore/TextDiff.swift
@@ -1,0 +1,1 @@
+// TextDiff public API is split across dedicated files.
