+++ Sources/TextDiffCore/TextDiffEngine.swift
