+++ Sources/TextDiffCore/TextDiffResult.swift
