+++ Sources/TextDiffCore/Tokenizer.swift
