+++ Sources/TextDiffMacOSUI/AppKit/DiffTextLayoutMetrics.swift
