+++ Sources/TextDiffMacOSUI/AppKit/NSTextDiffContentSource.swift
@@ -1,4 +1,5 @@
 import Foundation
+import TextDiffCore
 
 enum NSTextDiffContentSource {
     case text
