+++ Sources/TextDiffMacOSUI/TextDiffChangeStyle.swift
