+++ Sources/TextDiffMacOSUI/TextDiffChangeStyleDefaults.swift
