+++ Sources/TextDiffMacOSUI/TextDiffGroupStrokeStyle.swift
