+++ Sources/TextDiffMacOSUI/TextDiffStyle.swift
