+++ Sources/TextDiffMacOSUI/TextDiffStyling.swift
