+++ Sources/TextDiffMacOSUI/TextDiffViewModel.swift
@@ -1,5 +1,6 @@
 import Combine
 import Foundation
+import TextDiffCore
 
 @MainActor
 final class TextDiffViewModel: ObservableObject {
