+++ Tests/TextDiffMacOSUITests/DiffRevertActionResolverTests.swift
@@ -1,6 +1,7 @@
 import Foundation
 import Testing
-@testable import TextDiff
+import TextDiffCore
+@testable import TextDiffMacOSUI
 
 @Test
 func candidatesBuildPairedReplacementForAdjacentDeleteInsert() throws {
