+++ Tests/TextDiffMacOSUITests/NSTextDiffViewTests.swift
@@ -1,6 +1,7 @@
 import CoreGraphics
 import Testing
-@testable import TextDiff
+import TextDiffCore
+@testable import TextDiffMacOSUI
 
 @Test
 @MainActor
